requirements.txt
agnostic==1.0.2
asyncpg==0.21.0
black==20.8b1
discord.py==1.5.1
jishaku==1.19.2.202
lru-dict==1.1.6
pg8000==1.16.6
psutil==5.7.3
